image stuff

+++ src/image.rs
@@ -3,8 +3,8 @@ pub fn create_test_image(v: &mut Vec<u8>, width: u32, height: u32) {
     for i in (0..(height)).rev() {
         // println!("Lines remaining: {}", i + 1); //progress indicator in case of long renders
         for j in 0..width {
-            let r = i as f32 / (width - 1) as f32;
-            let g = j as f32 / (height - 1) as f32;
+            let r = i as f64 / (width - 1) as f64;
+            let g = j as f64 / (height - 1) as f64;
             let b = 0.25;
             // println!("{} {} {}", r, g, b);
